Don't rely on exceptions for testing assertion
The old way of turning asserts into exceptions for detection
made noexcept functions using asserts to not be testable as
noexcept called terminate when the assertion_failure exception
was thrown.

The new system use a function-level static variable to the error
message to be displayed. If none are ste, no assertions has been
triggered.